manual to other state position fix
Basically by telling the motor this is our constant for each position, we SHOULD be able to reset the positions at any state (without changing encoder constants), not just idle. This is unfinished, I'm unsure how exactly the parameter gathers the next state. Or in other words, where exactly the "nexState" is or should be defined. However, before I forget I wanted to commit my changes. If anyone can help or wants to work on it this is what I got.